refactor(web): DirectRow + deriveLinkLabel 를 primitives/DirectLinkRow 로 lift
BoldContactSidebar (Direct 섹션) 와 BoldProjectDetailHero (Links 섹션) 두
파일에 거의 동일한 DirectRow 컴포넌트 + URL host 라벨 추출 함수가 복사돼
있던 것을 단일 primitive 로 통합. 두 파일 모두 코멘트에 lift 검토 메모
존재했음.

- apps/web/components/primitives/DirectLinkRow.jsx 신설
  - DirectLinkRow ({ label, href, external = true, children })
  - deriveLinkLabel(url) named export (이전 deriveDirectLabel 도 같은 로직)
- BoldContactSidebar: 로컬 헬퍼 제거, primitive import. EMAIL 행은 external={false}
- BoldProjectDetailHero: 로컬 헬퍼 제거, primitive import

스타일 통합: gap-3 + truncate 적용 (Project 버전이 더 안전한 처리였음).

Closes #130
Refs #70
