+++ apps/web/components/projects/bold/detail/BoldProjectDetailHero.jsx
@@ -3,6 +3,9 @@ import Link from 'next/link';
 import Reveal from '../../../primitives/Reveal';
 import WordReveal from '../../../primitives/WordReveal';
 import Eyebrow from '../../../primitives/Eyebrow';
+import DirectLinkRow, {
+	deriveLinkLabel,
+} from '../../../primitives/DirectLinkRow';
 
 export default function BoldProjectDetailHero({
 	title,
@@ -118,8 +121,7 @@ export default function BoldProjectDetailHero({
 				</Reveal>
 			)}
 
-			{/* Links — admin 입력 외부 링크 (GitHub / Notion / Demo 등). 빈 배열이면 미렌더.
-			    Contact Sidebar 의 DirectRow 패턴 복사 (후속에 primitive lift 검토). */}
+			{/* Links — admin 입력 외부 링크 (GitHub / Notion / Demo 등). 빈 배열이면 미렌더. */}
 			{links.length > 0 && (
 				<Reveal
 					delay={0.28}
@@ -130,13 +132,13 @@ export default function BoldProjectDetailHero({
 					}}
 				>
 					{links.map((link) => (
-						<DirectRow
+						<DirectLinkRow
 							key={link.id ?? link.url}
 							label={deriveLinkLabel(link.url)}
 							href={link.url}
 						>
 							{link.label}
-						</DirectRow>
+						</DirectLinkRow>
 					))}
 				</Reveal>
 			)}
@@ -175,36 +177,6 @@ export default function BoldProjectDetailHero({
 // - accentWord 가 title 첫 토큰: accent / tail (2-line)
 // - accentWord 가 title 중간 토큰: head / accent / tail (3-line) — 시안 패턴
 // - accentWord 가 title 에 없거나 미지정: 마지막 토큰을 accent 로 폴백 (안전)
-// Contact Sidebar 의 DirectRow / deriveDirectLabel 패턴 그대로 복사 — 좌측 host
-// 라벨 자동 추출. 후속에 primitives/DirectLinkRow.jsx 로 lift 검토.
-function DirectRow({ label, href, children }) {
-	return (
-		<div className="flex items-center justify-between gap-3">
-			<span>{label}</span>
-			<a
-				href={href}
-				target="_blank"
-				rel="noopener noreferrer"
-				className="bold-interactive transition-colors hover:text-[color:var(--paper)] truncate"
-				style={{ color: 'var(--paper-dim)' }}
-			>
-				{children}
-			</a>
-		</div>
-	);
-}
-
-function deriveLinkLabel(url) {
-	try {
-		const host = new URL(url).host.toLowerCase();
-		const parts = host.split('.');
-		const root = parts.length >= 2 ? parts[parts.length - 2] : host;
-		return root.toUpperCase().slice(0, 10);
-	} catch {
-		return 'LINK';
-	}
-}
-
 function buildHeroItems(title, accentWord) {
 	const trimmed = (title ?? '').trim();
 	if (!trimmed) return [{ text: '' }];
